+++ src/core/MiniGames/ClickMoreMiniGame.ts
@@ -3,25 +3,27 @@ import { UserData } from '../../@types/UserTypes';
 import { TRANSLATION } from '../../lang/ru/translation';
 import MiniGame from '../MiniGame';
 
-import { Action } from './clickMoreMiniGame/Action';
+import { Player } from './clickMoreMiniGame/Player';
 
 type ClickMoreMiniGameProps = {
   players: Rivals;
   canvasId: string;
 };
 
-export type ActionProps = {
+export type BehaviorProps = {
   user: UserData;
   context: CanvasRenderingContext2D;
   canvas: HTMLCanvasElement;
   color: string;
-  leftSide: boolean;
+  isLeftSide: boolean;
+  addPointEvent: CustomEvent;
 };
 
 export class ClickMoreMiniGame extends MiniGame {
   finishCb: FinishFn | null;
   player: UserData | null;
   enemy: UserData | null;
+  addPointEvent: CustomEvent;
 
   constructor(props: ClickMoreMiniGameProps) {
     super({
@@ -32,10 +34,10 @@ export class ClickMoreMiniGame extends MiniGame {
     this.finishCb = () => null;
     this.player = null;
     this.enemy = null;
+    this.addPointEvent = new CustomEvent('addPointEvent', { bubbles: true });
   }
 
   run() {
-    console.log(this.players);
     return new Promise<MiniGameFinishResponse>(res => {
       this.finishCb = res;
       this.draw();
@@ -43,14 +45,15 @@ export class ClickMoreMiniGame extends MiniGame {
   }
 
   draw() {
-    const currentUser: ActionProps = {
+    const playerProps: BehaviorProps = {
       user: this.players[0],
       context: this.GameLoop.context,
       canvas: this.GameLoop.canvas,
       color: 'green',
-      leftSide: true,
+      isLeftSide: true,
+      addPointEvent: this.addPointEvent,
     };
-    new Action(currentUser).draw();
+    new Player(playerProps).draw();
   }
 
   finish(player: UserData) {
